QVTKOpenGLNativeWidget is used instead of QVTKWidget now.
This modification fixes the odd behaviors of Post2d window, Post2d bird's eye Window, and Post3d Window.

+++ libs/triangle/triangle.h
@@ -251,15 +251,15 @@
 #include "triangle_global.h"
 
 struct triangulateio {
-	REAL* pointlist;                                               /* In / out */
-	REAL* pointattributelist;                                      /* In / out */
+	double* pointlist;                                               /* In / out */
+	double* pointattributelist;                                      /* In / out */
 	int* pointmarkerlist;                                          /* In / out */
 	int numberofpoints;                                            /* In / out */
 	int numberofpointattributes;                                   /* In / out */
 
 	int* trianglelist;                                             /* In / out */
-	REAL* triangleattributelist;                                   /* In / out */
-	REAL* trianglearealist;                                         /* In only */
+	double* triangleattributelist;                                   /* In / out */
+	double* trianglearealist;                                         /* In only */
 	int* neighborlist;                                             /* Out only */
 	int numberoftriangles;                                         /* In / out */
 	int numberofcorners;                                           /* In / out */
@@ -269,15 +269,15 @@ struct triangulateio {
 	int* segmentmarkerlist;                                        /* In / out */
 	int numberofsegments;                                          /* In / out */
 
-	REAL* holelist;                        /* In / pointer to array copied out */
+	double* holelist;                        /* In / pointer to array copied out */
 	int numberofholes;                                      /* In / copied out */
 
-	REAL* regionlist;                      /* In / pointer to array copied out */
+	double* regionlist;                      /* In / pointer to array copied out */
 	int numberofregions;                                    /* In / copied out */
 
 	int* edgelist;                                                 /* Out only */
 	int* edgemarkerlist;            /* Not used with Voronoi diagram; out only */
-	REAL* normlist;                /* Used only with Voronoi diagram; out only */
+	double* normlist;                /* Used only with Voronoi diagram; out only */
 	int numberofedges;                                             /* Out only */
 };
